scripts/requirements.txt
elasticsearch>=8.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
